fix: フォント幅の取り扱いズレによる表示崩れの修正

+++ .config/home-manager/options/neovim/init.lua
@@ -49,7 +49,7 @@ opt.tabstop     = 2
 -- シフト幅
 opt.shiftwidth  = 2
 -- 全角文字
-opt.ambiwidth   = "double"
+opt.ambiwidth   = "single"
 -- シンタックスハイライト
 vim.cmd("syntax on")
 -- 対応するカッコやブレースを表示 + 時間
